xen: (Qubes) Allow kernel cmdline without kernel if stubdomains are used
I plan to fork Qemu's linux stubdomain to be a more generalized
stubdomain for servers. I don't know if this will be useful at all,
but since stubdomains are PV VMs this might actually be useful for us.
As far as I can tell stubdomains have to be booted using direct kernel
boot. This might be related to libvirt compatibility, in which case it
is useful to me.