xen: (Qubes) Allow up to 4 xenconsole instances per VM
Normally a VM only gets a single xenconsole assigned to it and the rest
are provided by QEMU. In order to get rid of QEMU being used in dom0 for
the consoles we instead extend xenconsoled to support 4 consoles per VM.
I think this is a good compromise between security and usability. Do
note that when stubdomains are used some of the consoles will be used up
by the stubdomains.